Extract VaultDetailsCardCollaterlizationRatio.
